requirements.txt
# requirements.txt for YouTube Downloader Pro v2.0

# Core — YouTube downloading
yt-dlp>=2023.10.13

# Image processing for thumbnails
pillow>=10.0.0

# HTTP requests for thumbnails
requests>=2.31.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0